hwcrypto: split the key derive implementation with platforms
