+++ hwcrypto/hwkey_srv_provider.c
@@ -45,52 +45,27 @@
 
 #define CRYPTO_CTX_INITIAL_VALUE(crypto_ctx) {0, {0}, {0}, {0}, {0}, 0, {0}}
 
-static struct crypto_context g_crypto_ctx = CRYPTO_CTX_INITIAL_VALUE(g_crypto_ctx);
+struct crypto_context g_crypto_ctx = CRYPTO_CTX_INITIAL_VALUE(g_crypto_ctx);
 
-struct key {
-	uint8_t byte[32];
-};
-
-struct iv {
-	uint8_t byte[12];
-};
-
-struct aad {
-	uint8_t byte[16];
-};
-
-struct tag {
-	uint8_t byte[16];
-};
-
-#define AES_GCM_NO_ERROR           0
-#define AES_GCM_ERR_GENERIC        -1
-#define AES_GCM_ERR_AUTH_FAILED    -2
-
-static const struct aad trk_aad = {
+const struct aad trk_aad = {
 	.byte = {
 		0xf3, 0x56, 0x5b, 0xd9, 0xc4, 0xe7, 0xd4, 0x1e,
 		0xbb, 0xb4, 0x14, 0x15, 0x20, 0xe7, 0x09, 0xcf,
 	}
 };
 
-static const struct aad ssek_aad = {
+const struct aad ssek_aad = {
 	.byte = {
 		0x8d, 0x46, 0x2b, 0xd1, 0xb3, 0xde, 0x0f, 0x5c,
 		0xc1, 0x6d, 0x56, 0xcc, 0x2e, 0x53, 0x05, 0x54,
 	}
 };
+/* Secure storage service app uuid */
+const uuid_t ss_uuid = SECURE_STORAGE_SERVER_APP_UUID;
+const uuid_t crypto_uuid = HWCRYPTO_SRV_APP_UUID;
 
 static const uint8_t gcm_info[] = {0xc7, 0x87, 0xaf, 0xe7, 0x3b, 0xca, 0x44, 0x63,
 				   0x35, 0x16, 0x0b, 0x94, 0x52, 0x53, 0x4d, 0xa3};
-
-/* RPMB Key support */
-#define RPMB_SS_AUTH_KEY_SIZE    32
-#define RPMB_SS_AUTH_KEY_ID      "com.android.trusty.storage_auth.rpmb"
-
-/* Secure storage service app uuid */
-static const uuid_t ss_uuid = SECURE_STORAGE_SERVER_APP_UUID;
-static const uuid_t crypto_uuid = HWCRYPTO_SRV_APP_UUID;
 
 #if LK_DEBUGLEVEL > 1
 
@@ -172,7 +147,7 @@ static bool hwkey_self_test(void)
  *
  * Return: 0 on success, < 0 if an error was detected.
  */
-static int aes_256_gcm_encrypt(const struct key *key,
+int aes_256_gcm_encrypt(const struct key *key,
 			const struct iv *iv, const struct aad *aad,
 			const void *plain, size_t plain_size,
 			void *out, size_t *out_size)
@@ -282,7 +257,7 @@ exit:
  *
  * Return: 0 on success, < 0 if an error was detected.
  */
-static int aes_256_gcm_decrypt(const struct key *key,
+int aes_256_gcm_decrypt(const struct key *key,
 			const struct iv *iv, const struct aad *aad,
 			const void *cipher, size_t cipher_size,
 			void *out, size_t *out_size)
@@ -374,7 +349,7 @@ exit:
 	return rc;
 }
 
-static int get_device_index_huk(uint8_t index, uint8_t *huk, uint32_t huk_len)
+static int get_device_seed_by_index(uint8_t index, uint8_t *seed, uint32_t seed_len)
 {
 	int rc = 0;
 	trusty_device_info_t dev_info;
@@ -382,7 +357,7 @@ static int get_device_index_huk(uint8_t index, uint8_t *huk, uint32_t huk_len)
 	if (index >= CSE_SEED_MAX_ENTRIES)
 		return HWKEY_ERR_NOT_VALID;
 
-	if(!huk) {
+	if(!seed) {
 		TLOGE("the input param is NULL\n");
 		return ERR_IO;
 	}
@@ -407,7 +382,7 @@ static int get_device_index_huk(uint8_t index, uint8_t *huk, uint32_t huk_len)
 	 * will be used for now by Trusty. But keep higher 32 bytes for future extension.
 	 */
 
-	rc = memcpy_s(huk, huk_len, dev_info.sec_info.dseed_list[index].seed, huk_len);
+	rc = memcpy_s(seed, seed_len, dev_info.sec_info.dseed_list[index].seed, seed_len);
 
 clear_sensitive_data:
 	secure_memzero(&dev_info, sizeof(trusty_device_info_t));
@@ -460,14 +435,14 @@ static uint8_t get_svn_by_index(uint8_t index)
 }
 
 /* aes gcm key for crypto is derived from seed[index] */
-static uint32_t get_aes_gcm_key(uint8_t index, uint8_t *aes_gcm_key, size_t key_len)
+uint32_t get_aes_gcm_key(uint8_t index, uint8_t *aes_gcm_key, size_t key_len)
 {
 	int rc = -1;
 	uint8_t hw_device_key[32] = {0};
 
 	assert(aes_gcm_key);
 
-	if (get_device_index_huk(index, hw_device_key, sizeof(hw_device_key))) {
+	if (get_device_seed_by_index(index, hw_device_key, sizeof(hw_device_key))) {
 		TLOGE("failed (%d) to get device HUK.\n", rc);
 		goto out;
 	}
@@ -689,227 +664,6 @@ out:
 	return rc;
 }
 
-uint32_t get_ssek(uint8_t *ssek, size_t *ssek_len)
-{
-	uint8_t aes_gcm_key[32] = {0};
-	int rc = -1;
-
-	assert(ssek && ssek_len);
-
-	/* always use seed[0] derivative to decrypt ssek and trk */
-	rc = get_aes_gcm_key(0, aes_gcm_key, sizeof(aes_gcm_key));
-	if (rc != NO_ERROR) {
-		*ssek_len = 0;
-		TLOGE("get_ssek failed (%d) to get_aes_gcm_key.\n", rc);
-		goto out;
-	}
-
-	rc = aes_256_gcm_decrypt((const struct key *)aes_gcm_key,
-				(const struct iv *)g_crypto_ctx.ssek_iv, &ssek_aad,
-				(const void *)g_crypto_ctx.ssek_cipher, sizeof(g_crypto_ctx.ssek_cipher),
-				ssek, ssek_len);
-	if (rc || *ssek_len != sizeof(struct key)) {
-		TLOGE("get_ssek failed to decrypt ssek, rc is %d. *ssek_len is %zu.\n", rc, *ssek_len);
-		*ssek_len = 0;
-		secure_memzero(ssek, sizeof(struct key));
-		goto out;
-	}
-
-	rc = HWKEY_NO_ERROR;
-out:
-	secure_memzero(aes_gcm_key, sizeof(aes_gcm_key));
-	return rc;
-}
-
-/*
- * Derive key V1 - HMAC SHA256 based Key derivation function
- */
-#ifdef SIMICS_HARDCODE_KEY
-uint32_t derive_key_v1(const uuid_t *uuid,
-			const uint8_t *ikm_data, size_t ikm_len,
-			uint8_t *key_buf, size_t *key_len)
-{
-	*key_len = ikm_len;
-	memset(key_buf, 0, ikm_len);
-
-	return HWKEY_NO_ERROR;
-}
-#else
-uint32_t derive_key_v1(const uuid_t *uuid,
-			const uint8_t *ikm_data, size_t ikm_len,
-			uint8_t *key_buf, size_t *key_len)
-{
-
-	struct key trk;
-	size_t out_size = 0;
-	uint8_t aes_gcm_key[32] = {0};
-	int rc = -1;
-
-	assert(ikm_data && key_buf && key_buf);
-
-	if (!ikm_len) {
-		*key_len = 0;
-		return HWKEY_ERR_BAD_LEN;
-	}
-
-	/* always use seed 0 derivative to decrypt ssek and trk */
-	rc = get_aes_gcm_key(0, aes_gcm_key, sizeof(aes_gcm_key));
-	if (rc != NO_ERROR) {
-		TLOGE("failed (%d) to get device HUK\n", rc);
-		goto out;
-	}
-
-	rc = aes_256_gcm_decrypt((const struct key *)aes_gcm_key,
-				(const struct iv *)g_crypto_ctx.trk_iv, &trk_aad,
-				(const void *)g_crypto_ctx.trk_cipher, sizeof(g_crypto_ctx.trk_cipher),
-				&trk, &out_size);
-	if (rc || out_size != sizeof(trk)) {
-		TLOGE("aes_256_gcm_decrypt failed to decrypt ssek, rc is %d. out_size is %zu.\n", rc, out_size);
-		*key_len = 0;
-		secure_memzero(key_buf, ikm_len);
-		goto out;
-        }
-
-	if (!HKDF(key_buf, ikm_len, EVP_sha256(),
-		(const uint8_t *)&trk, sizeof(trk),
-		(const uint8_t *)uuid, sizeof(uuid_t),
-		ikm_data, ikm_len)) {
-		TLOGE(" derive_key_v1 HDKF failed 0x%x.\n", ERR_get_error());
-		*key_len = 0;
-		secure_memzero(key_buf, ikm_len);
-		goto out;
-	}
-
-	*key_len = ikm_len;
-	rc = HWKEY_NO_ERROR;
-
-out:
-	secure_memzero(aes_gcm_key, sizeof(aes_gcm_key));
-	secure_memzero(&trk, sizeof(trk));
-
-	return rc;
-}
-#endif
-
-static int get_device_huk(uint8_t *huk, uint32_t huk_len)
-{
-	return get_device_index_huk(0, huk, huk_len);
-}
-
-/*
- * Generate RPMB Secure Storage Authentication key from seed[index]
- */
-static uint32_t get_rpmb_ss_auth_key_with_index(uint8_t index,
-				uint8_t *kbuf, size_t kbuf_len, size_t *klen)
-{
-	uint8_t rpmb_key[RPMB_SS_AUTH_KEY_SIZE] = {0};
-	trusty_device_info_t dev_info;
-	int ret = HWKEY_ERR_GENERIC;
-	uint8_t serial[MMC_PROD_NAME_WITH_PSN_LEN] = {0};
-
-	assert(kbuf);
-	assert(klen);
-
-	if (NO_ERROR != get_device_info(&dev_info)) {
-		TLOGE("failed to get device infomation\n");
-		goto out;
-	}
-
-	if (index < dev_info.sec_info.num_seeds) {
-		/* Clear Byte 2 and 0 for CID[6] PRV and CID[0] CRC for eMMC Field Firmware Updates
-		 * serial[0] = cid[0];  -- CRC
-		 * serial[2] = cid[6];  -- PRV
-		 */
-		memcpy_s(serial, sizeof(serial), dev_info.sec_info.serial, sizeof(dev_info.sec_info.serial));
-		serial[0] ^= serial[0];
-		serial[2] ^= serial[2];
-
-		if (!HKDF(rpmb_key, sizeof(rpmb_key), EVP_sha256(),
-			  (const uint8_t *)dev_info.sec_info.dseed_list[index].seed, 32,
-			  (const uint8_t *)&crypto_uuid, sizeof(uuid_t),
-			  (const uint8_t *)serial, sizeof(serial))) {
-			TLOGE("HDKF failed 0x%x\n", ERR_get_error());
-			goto out;
-		}
-		memcpy_s(kbuf, RPMB_SS_AUTH_KEY_SIZE, rpmb_key, RPMB_SS_AUTH_KEY_SIZE);
-	}
-	*klen = RPMB_SS_AUTH_KEY_SIZE;
-
-	ret = HWKEY_NO_ERROR;
-out:
-	secure_memzero(rpmb_key, sizeof(rpmb_key));
-	secure_memzero(&dev_info, sizeof(dev_info));
-
-	return ret;
-}
-
-/*
- * Generate RPMB Secure Storage Authentication keys for all seeds
- */
-static uint32_t get_rpmb_ss_auth_key(const struct hwkey_keyslot *slot,
-				     uint8_t *kbuf, size_t kbuf_len, size_t *klen)
-{
-	int rc = 0;
-	uint32_t i;
-	trusty_device_info_t dev_info;
-	size_t klen_for_once = 0;
-	uint8_t invlid_key[64] = {0};
-
-	assert(kbuf);
-	assert(klen);
-
-	*klen = 0;
-
-	rc = get_device_info(&dev_info);
-	if (NO_ERROR != rc) {
-		TLOGE("%s:failed to get device infomation\n", __func__);
-		rc = HWKEY_ERR_GENERIC;
-		goto clear_dev_info;
-	}
-
-	switch (dev_info.sec_info.platform)
-	{
-		case APL_PLATFORM:
-			for (i = 0; i < dev_info.sec_info.num_seeds; i++) {
-				if (HWKEY_NO_ERROR != get_rpmb_ss_auth_key_with_index(
-						i, kbuf + i * RPMB_SS_AUTH_KEY_SIZE, kbuf_len, &klen_for_once)) {
-					secure_memzero(kbuf, kbuf_len);
-					rc = HWKEY_ERR_GENERIC;
-					goto clear_dev_info;
-				}
-				*klen += klen_for_once;
-			}
-			break;
-
-		case ICL_PLATFORM:
-			if (!CRYPTO_memcmp(dev_info.sec_info.rpmb_key[0], invlid_key, sizeof(invlid_key)))
-			{
-				TLOGE("%s: the RPMB key is unavailable.\n", __func__);
-				rc = HWKEY_ERR_GENERIC;
-				goto clear_dev_info;
-			}
-
-			memcpy_s(kbuf, RPMB_SS_AUTH_KEY_SIZE, dev_info.sec_info.rpmb_key[0], RPMB_SS_AUTH_KEY_SIZE);
-			*klen = RPMB_SS_AUTH_KEY_SIZE;
-			break;
-
-		case DUMMY_PLATFORM:
-			memcpy_s(kbuf, RPMB_SS_AUTH_KEY_SIZE, "12345ABCDEF1234512345ABCDEF12345", RPMB_SS_AUTH_KEY_SIZE);
-			*klen = RPMB_SS_AUTH_KEY_SIZE;
-			break;
-
-		default:
-			//TODO: CWP rpmb key.
-			TLOGE("%s: platform(%d) is not handled!\n", __func__, dev_info.sec_info.platform);
-			assert(0);
-			break;
-	}
-
-clear_dev_info:
-	secure_memzero(&dev_info, sizeof(trusty_device_info_t));
-	return rc;
-}
-
 /*
  *  List of keys slots that hwkey service supports
  */
@@ -927,9 +681,7 @@ static const struct hwkey_keyslot _keys[] = {
 void hwkey_init_srv_provider(void)
 {
 	int rc;
-#ifndef SIMICS_HARDCODE_KEY
 	uint32_t seed_count, i;
-#endif
 
 	TLOGI("Init hwkey service provider\n");
 
@@ -951,7 +703,6 @@ void hwkey_init_srv_provider(void)
 		abort();
 	}
 
-#ifndef SIMICS_HARDCODE_KEY
 	if (get_seed_count(&seed_count))
 		abort();
 
@@ -962,6 +713,5 @@ void hwkey_init_srv_provider(void)
 			abort();
 		}
 	}
-#endif
 }
